Moving config update related functions to a separate file
