+++ pkg/resources/kafka/kafka.go
@@ -24,7 +24,6 @@ import (
 	"time"
 
 	"emperror.dev/errors"
-	"github.com/Shopify/sarama"
 	"github.com/banzaicloud/k8s-objectmatcher/patch"
 	"github.com/go-logr/logr"
 	corev1 "k8s.io/api/core/v1"
@@ -489,136 +488,6 @@ func (r *Reconciler) getServerAndClientDetails() (string, string, []string, erro
 	return serverPass, clientPass, superUsers, nil
 }
 
-func (r *Reconciler) reconcilePerBrokerDynamicConfig(brokerId int32, brokerConfig *v1beta1.BrokerConfig, configMap *corev1.ConfigMap, log logr.Logger) error {
-	kClient, err := r.kafkaClientProvider.NewFromCluster(r.Client, r.KafkaCluster)
-	if err != nil {
-		return errorfactory.New(errorfactory.BrokersUnreachable{}, err, "could not connect to kafka brokers")
-	}
-	defer func() {
-		if err := kClient.Close(); err != nil {
-			log.Error(err, "could not close client")
-		}
-	}()
-
-	parsedBrokerConfig := util.ParsePropertiesFormat(brokerConfig.Config)
-
-	currentPerBrokerConfigState := r.KafkaCluster.Status.BrokersState[strconv.Itoa(int(brokerId))].PerBrokerConfigurationState
-	if len(parsedBrokerConfig) == 0 && currentPerBrokerConfigState != v1beta1.PerBrokerConfigOutOfSync {
-		return nil
-	}
-
-	// overwrite configs that are generated in the configmap
-	configsFromConfigMap := k8sutil.GetBrokerConfigsFromConfigMap(configMap)
-	for _, perBrokerConfig := range k8sutil.PerBrokerConfigs {
-		if configValue, ok := configsFromConfigMap[perBrokerConfig]; ok {
-			parsedBrokerConfig[perBrokerConfig] = configValue
-		}
-	}
-
-	brokerConfigKeys := make([]string, 0, len(parsedBrokerConfig)+len(k8sutil.PerBrokerConfigs))
-	for key := range parsedBrokerConfig {
-		brokerConfigKeys = append(brokerConfigKeys, key)
-	}
-	brokerConfigKeys = append(brokerConfigKeys, k8sutil.PerBrokerConfigs...)
-
-	response, err := kClient.DescribePerBrokerConfig(brokerId, brokerConfigKeys)
-	if err != nil {
-		return errors.WrapIfWithDetails(err, "could not describe broker config", "brokerId", brokerId)
-	}
-
-	changedConfigs := collectNonIdenticalConfigsFromResponse(response, parsedBrokerConfig)
-
-	if len(changedConfigs) > 0 {
-		if currentPerBrokerConfigState == v1beta1.PerBrokerConfigInSync {
-			log.V(1).Info("setting per broker config status to out of sync")
-			statusErr := k8sutil.UpdateBrokerStatus(r.Client, []string{strconv.Itoa(int(brokerId))}, r.KafkaCluster, v1beta1.PerBrokerConfigOutOfSync, log)
-			if statusErr != nil {
-				return errors.WrapIfWithDetails(err, "updating status for per-broker configuration status failed", "brokerId", brokerId)
-			}
-		}
-
-		// validate the config
-		err := kClient.AlterPerBrokerConfig(brokerId, util.ConvertMapStringToMapStringPointer(changedConfigs), true)
-		if err != nil {
-			statusErr := k8sutil.UpdateBrokerStatus(r.Client, []string{strconv.Itoa(int(brokerId))}, r.KafkaCluster, v1beta1.PerBrokerConfigError, log)
-			if statusErr != nil {
-				err = errors.Combine(err, statusErr)
-			}
-			return errors.WrapIfWithDetails(err, "could not validate per-broker broker config", "brokerId", brokerId)
-		}
-
-		// alter the config
-		err = kClient.AlterPerBrokerConfig(brokerId, util.ConvertMapStringToMapStringPointer(changedConfigs), false)
-		if err != nil {
-			return errors.WrapIfWithDetails(err, "could not alter broker config", "brokerId", brokerId)
-		}
-
-		// request for the updated config
-		response, err := kClient.DescribePerBrokerConfig(brokerId, brokerConfigKeys)
-		if err != nil {
-			return errors.WrapIfWithDetails(err, "could not describe broker config", "brokerId", brokerId)
-		}
-
-		// update per-broker config status
-		notUpdatedConfigs := collectNonIdenticalConfigsFromResponse(response, parsedBrokerConfig)
-		if len(notUpdatedConfigs) > 0 {
-			return errorfactory.New(errorfactory.PerBrokerConfigNotReady{}, errors.New("configuration is out of sync"), "per-broker configuration updated")
-		} else {
-			statusErr := k8sutil.UpdateBrokerStatus(r.Client, []string{strconv.Itoa(int(brokerId))}, r.KafkaCluster, v1beta1.PerBrokerConfigInSync, log)
-			if statusErr != nil {
-				return errors.WrapIfWithDetails(err, "updating status for per-broker configuration status failed", "brokerId", brokerId)
-			}
-		}
-	} else {
-		log.V(1).Info("setting per broker config status to in sync")
-		statusErr := k8sutil.UpdateBrokerStatus(r.Client, []string{strconv.Itoa(int(brokerId))}, r.KafkaCluster, v1beta1.PerBrokerConfigInSync, log)
-		if statusErr != nil {
-			return errors.WrapIfWithDetails(err, "updating status for per-broker configuration status failed", "brokerId", brokerId)
-		}
-	}
-
-	return nil
-}
-
-func (r *Reconciler) reconcileClusterWideDynamicConfig(log logr.Logger) error {
-	kClient, err := r.kafkaClientProvider.NewFromCluster(r.Client, r.KafkaCluster)
-	if err != nil {
-		return errorfactory.New(errorfactory.BrokersUnreachable{}, err, "could not connect to kafka brokers")
-	}
-	defer func() {
-		if err := kClient.Close(); err != nil {
-			log.Error(err, "could not close client")
-		}
-	}()
-
-	configIdentical := true
-
-	currentConfig, err := kClient.DescribeClusterWideConfig()
-	if err != nil {
-		return errors.WrapIf(err, "could not describe cluster wide broker config")
-	}
-	parsedClusterWideConfig := util.ParsePropertiesFormat(r.KafkaCluster.Spec.ClusterWideConfig)
-	if len(currentConfig) != len(parsedClusterWideConfig) {
-		configIdentical = false
-	}
-	for _, conf := range currentConfig {
-		if val, ok := parsedClusterWideConfig[conf.Name]; ok {
-			if val != conf.Value {
-				configIdentical = false
-				break
-			}
-		}
-	}
-	if !configIdentical {
-		err = kClient.AlterClusterWideConfig(util.ConvertMapStringToMapStringPointer(parsedClusterWideConfig))
-		if err != nil {
-			return errors.WrapIf(err, "could not alter cluster wide broker config")
-		}
-	}
-
-	return nil
-}
-
 func (r *Reconciler) reconcileKafkaPod(log logr.Logger, desiredPod *corev1.Pod) error {
 	currentPod := desiredPod.DeepCopy()
 	desiredType := reflect.TypeOf(desiredPod)
@@ -934,17 +803,4 @@ func GetBrokersWithPendingOrRunningCCTask(kafkaCluster *v1beta1.KafkaCluster) []
 
 func isDesiredStorageValueInvalid(desired, current *corev1.PersistentVolumeClaim) bool {
 	return desired.Spec.Resources.Requests.Storage().Value() < current.Spec.Resources.Requests.Storage().Value()
-}
-
-func collectNonIdenticalConfigsFromResponse(response []*sarama.ConfigEntry, parsedBrokerConfig map[string]string) map[string]string {
-	changedConfigs := make(map[string]string, 0)
-	for _, conf := range response {
-		if val, ok := parsedBrokerConfig[conf.Name]; ok {
-			if val != conf.Value {
-				changedConfigs[conf.Name] = val
-			}
-		}
-	}
-
-	return changedConfigs
 }
